+++ .gitignore
@@ -15,6 +15,8 @@ data/*.csv
 # FOR NOW, until we verify the redaction on these pages:
 test/mock_dmap/*.mhtml
 test/mock_dmap/*.html
+#!test/mock_dmap/dashboard-1-redacted.mhtml
+#!test/mock_dmap/dashboard-2-redacted.mhtml
 
 exports/*.csv
 exports/*/*.png
